+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.3.9 (2024-04-xx)
+### Common
+* **core**: `GET /detections` return `Total-items` in response headers
+
+### Bug Fixes
+* **core**: Fix update only stream name also set default to timezone and country code
+
 ## 1.3.8 (2024-03-xx)
 ### Common
 * **core**: Use Google Map API for getting timezone and country from latitude and longitude